requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
pandas
joblib
scikit-learn
faiss-cpu
sentence-transformers
requests
beautifulsoup4
pypdf
mangum
